Implement the negamax based on cmwq
Implement execution of two different tic-tac-toe AI algorithms (negamax) on separate CPUs and simulate their gameplay. Utilize CMWQ for asynchronous execution.

+++ game.h
@@ -25,5 +25,5 @@ extern const line_t lines[4];
 
 int *available_moves(char *table);
 char check_win(char *t);
-double calculate_win_value(char win, char player);
+int calculate_win_value(char win, char player);
 // static int draw_board(char *t);
